refactor: align station insights component naming
