+++ packages/frontend-next/src/components/map/StationHistoricalInsights.tsx
